requirements/requrements.txt
APScheduler==3.3.1
baidupush==0.1.1
bottle==0.12.13
certifi==2017.4.17
chardet==3.0.3
Django==1.11.3
django-cors-headers==2.1.0
django-debug-toolbar==1.8
django-extra-fields==0.9
django-filter==1.0.4
djangorestframework==3.6.3
gunicorn==19.7.1
idna==2.5
Markdown==2.6.8
MarkdownSuperscript==2.0.0
MarkupSafe==1.0
olefile==0.44
pika==0.10.0
Pillow==4.1.1
pymongo==3.4.0
PyMySQL==0.7.11
pytz==2017.2
requests==2.17.3
six==1.10.0
sqlparse==0.2.3
tornado==4.5.1
tzlocal==1.4
ujson==1.35
urllib3==1.21.1
WeRoBot==1.1.1
xmltodict==0.11.0